backend/requirements.txt
agno==2.2.5
annotated-types==0.7.0
anyio==4.11.0
arxiv==2.2.0
asgiref==3.10.0
beautifulsoup4==4.14.2
brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==46.0.3
ddgs==9.6.1
django==5.2.7
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-ninja==1.4.5
django-ninja-extra==0.30.2
django-ninja-jwt==5.4.0
django-redis==6.0.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
feedparser==6.0.12
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.42.1
google-genai==1.47.0
googlesearch-python==1.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
injector==0.22.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pillow==12.0.0
pip==25.0.1
primp==0.15.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
pypdf==6.1.3
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
redis==7.0.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
sgmllib3k==1.0.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sqlparse==0.5.3
tenacity==9.1.2
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
